requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=hd4dd3e8_0
absl-py=2.0.0=pypi_0
aiohttp=3.8.5=py310h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
albumentations=1.3.1=pypi_0
appdirs=1.4.4=pypi_0
argparse=1.4.0=pypi_0
arrow-cpp=11.0.0=h374c478_2
asttokens=2.4.1=pyhd8ed1ab_0
async-timeout=4.0.2=py310h06a4308_0
attrs=21.4.0=pypi_0
av=10.0.0=pypi_0
aws-c-common=0.6.8=h5eee18b_1
aws-c-event-stream=0.1.6=h6a678d5_6
aws-checksums=0.1.11=h5eee18b_2
aws-sdk-cpp=1.8.185=h721c034_1
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
blas=1.0=mkl
boost-cpp=1.82.0=hdb19cb5_2
bottleneck=1.3.5=py310ha9d4c09_0
brotli-python=1.0.9=py310h6a678d5_7
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.12.12=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2022.12.7=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=2.1.1=pypi_0
click=8.1.7=pypi_0
cmake=3.25.0=pypi_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.1.1=pypi_0
cryptography=37.0.4=pypi_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.3.101=0
cuda-runtime=12.1.0=0
cycler=0.12.1=pypi_0
datasets=2.12.0=py310h06a4308_0
debugpy=1.6.7=py310h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.6=py310h06a4308_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py310h06a4308_0
fonttools=4.43.1=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=py310h5eee18b_0
fsspec=2023.9.2=py310h06a4308_0
fvcore=0.1.5.post20221221=pypi_0
gdrive=0.1.5=pypi_0
gflags=2.2.2=he6710b0_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
glog=0.5.0=h2531618_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-api-core=2.15.0=pypi_0
google-api-python-client=2.43.0=pypi_0
google-auth=2.25.2=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-auth-oauthlib=0.5.3=pypi_0
googleapis-common-protos=1.62.0=pypi_0
grpc-cpp=1.48.2=he1ff14a_1
grpcio=1.60.0=pypi_0
httplib2=0.22.0=pypi_0
huggingface_hub=0.17.3=py310h06a4308_0
icu=73.1=h6a678d5_0
idna=3.4=py310h06a4308_0
imageio=2.31.6=pypi_0
importlib-metadata=6.0.0=py310h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
iopath=0.1.10=pypi_0
ipykernel=6.26.0=pyhf8b6a83_0
ipympl=0.9.3=pypi_0
ipython=8.16.1=pyh0d859eb_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.1.1=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jeepney=0.8.0=pypi_0
jinja2=3.1.2=py310h06a4308_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.4.0=py310hff52083_0
jupyterlab-widgets=3.0.9=pypi_0
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libboost=1.82.0=h109eef0_2
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.8.1.2=0
libcurand=10.3.4.101=0
libcurl=8.4.0=h251f7ec_0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnghttp2=1.57.0=h2d74bed_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h36c2ea0_1
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libthrift=0.15.0=h1795dd8_2
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
lightning-utilities=0.9.0=pypi_0
lit=15.0.7=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.5.1=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.8.0=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.8=py310h5eee18b_0
mkl_random=1.2.4=py310hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
multidict=6.0.2=py310h5eee18b_0
multiprocess=0.70.14=py310h06a4308_0
natsort=8.4.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.0=pypi_0
numexpr=2.8.7=py310h85018f9_0
numpy=1.26.1=pypi_0
numpy-base=1.26.0=py310hb5e798b_0
oauthlib=3.2.2=pypi_0
opencv-python=4.8.1.78=pypi_0
opencv-python-headless=4.8.1.78=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.12=h7f8727e_0
orc=1.7.4=hb3bc3d3_1
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.1=py310h1128e8f_0
parameterized=0.9.0=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.3.0=pypi_0
pip=23.3=py310h06a4308_0
platformdirs=3.11.0=pyhd8ed1ab_0
portalocker=2.8.2=pypi_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
protobuf=4.25.2=pypi_0
psutil=5.9.6=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=11.0.0=py310h468efa6_1
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycocotools=2.0.7=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.16.1=pyhd8ed1ab_0
pyopenssl=23.2.0=py310h06a4308_0
pyparsing=3.1.1=pypi_0
pyqt6=6.6.0=pypi_0
pyqt6-qt6=6.6.0=pypi_0
pyqt6-sip=13.6.0=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.12=h955ad1f_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py310h5eee18b_1
python_abi=3.10=2_cp310
pytomlpp=1.0.13=pypi_0
pytorch=2.1.2=py3.10_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytorchvideo=0.1.5=pypi_0
pytz=2023.3.post1=py310h06a4308_0
pyyaml=5.4.1=pypi_0
pyzmq=25.1.0=py310h6a678d5_0
qudida=0.0.4=pypi_0
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
regex=2023.10.3=py310h5eee18b_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
responses=0.13.3=pyhd3eb1b0_0
rsa=4.9=pypi_0
safetensors=0.4.0=py310ha89cbab_0
scikit-image=0.22.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.3=pypi_0
seaborn=0.13.1=pypi_0
secretstorage=3.3.3=pypi_0
sentry-sdk=1.32.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=59.6.0=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
snappy=1.1.9=h295c915_0
spock-config=3.0.2=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=py310h06a4308_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.15.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=2.3.0=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.2.0=pypi_0
tifffile=2023.9.26=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=py310h22610ee_0
torch=2.0.0+cu118=pypi_0
torch-tb-profiler=0.4.3=pypi_0
torchaudio=2.0.1+cu118=pypi_0
torchmetrics=1.2.0=pypi_0
torchtriton=2.1.0=py310
torchvision=0.15.1+cu118=pypi_0
tornado=6.1=py310h5764c6d_3
tqdm=4.66.1=pypi_0
traitlets=5.12.0=pyhd8ed1ab_0
transformers=4.32.1=py310h06a4308_0
triton=2.0.0=pypi_0
typing-extensions=4.4.0=pypi_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023c=h04d1e81_0
uritemplate=4.1.1=pypi_0
urllib3=1.26.13=pypi_0
utf8proc=2.6.1=h27cfd23_0
versioneer=0.29=pypi_0
wandb=0.15.12=pypi_0
wcwidth=0.2.8=pyhd8ed1ab_0
werkzeug=3.0.1=pypi_0
wheel=0.37.1=pypi_0
widgetsnbextension=4.0.9=pypi_0
xxhash=0.8.0=h7f8727e_3
xz=5.4.2=h5eee18b_0
yacs=0.1.8=pypi_0
yaml=0.2.5=h7b6447c_0
yarl=1.8.1=py310h5eee18b_0
yolov7detect=1.0.1=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.11.0=py310h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
